Refactor: Update data type of TotalRevenue in wac_activities table to DECIMAL(19, 4) Refactor: Rename column in wac_activities table to wac_id Refactor: Update TotalRevenue data type in wac_activities repository to float64

+++ db/migrations/20240925153824_create_wac_activities_table.up.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS wac_activities (
     total_potential_leads INT NOT NULL DEFAULT 0,
     total_leads INT NOT NULL DEFAULT 0,
     total_completed_leads INT NOT NULL DEFAULT 0,
-    total_revenue INT NOT NULL DEFAULT 0,
+    total_revenue DECIMAL(19, 4) NOT NULL DEFAULT 0.0000,
 
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now(),
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now(),
